feat: moving the string identifier to its own file
